fix(coding-agent): stop goal continuation after terminal turns
