+++ packages/coding-agent/src/core/extensions/builtin/goal/index.ts
@@ -1,28 +1,23 @@
 import { createHash } from "node:crypto";
 import { join } from "node:path";
-import { Type } from "typebox";
 import { getAgentDir } from "../../../../config.ts";
-import type { AgentToolResult, ExtensionAPI, ExtensionContext } from "../../types.ts";
-import { parseGoalCommand } from "./command.ts";
+import type { ExtensionAPI, ExtensionContext } from "../../types.ts";
+import { registerGoalCommand } from "./command-registration.ts";
 import { shouldQueueGoalContinuationAfterAgentEnd, shouldQueueGoalContinuationWhenIdle } from "./continuation.ts";
-import { formatGoalForTool, formatGoalToolResponse, goalStatusLabel } from "./format.ts";
+import { formatGoalForTool, goalStatusLabel } from "./format.ts";
 import { buildContinuationPrompt } from "./prompt.ts";
-import { accountGoalUsage, clearGoal, createGoal, readGoal, updateGoal } from "./store.ts";
+import { accountGoalUsage, readGoal, updateGoal } from "./store.ts";
+import { registerGoalTools } from "./tool-registration.ts";
 import type { Goal, GoalAccountingMode, GoalStoreRef, TokenUsageSnapshot } from "./types.ts";
-import { COMPLETABLE_GOAL_STATUS_VALUES, isRecord } from "./types.ts";
+import { isRecord } from "./types.ts";
 import { updateGoalUi } from "./ui.ts";
 
-const GOAL_USAGE = "Usage: /goal <objective>";
-const GOAL_EMPTY_HINT = "No goal is currently set.";
 const GOAL_CONTINUATION_MESSAGE_TYPE = "goal-continuation";
-const REPLACE_GOAL_CHOICE = "Replace current goal";
-const CANCEL_REPLACE_GOAL_CHOICE = "Cancel";
 const RESUME_GOAL_CHOICE = "Resume goal";
 const LEAVE_GOAL_PAUSED_CHOICE = "Leave paused";
 const EMPTY_USAGE: TokenUsageSnapshot = { input: 0, output: 0, cacheRead: 0, cacheWrite: 0, totalTokens: 0 };
 const STALE_EXTENSION_CONTEXT_ERROR_PREFIX = "This extension ctx is stale after session replacement or reload.";
 
-type GoalToolResult = AgentToolResult<Record<string, never>>;
 type AssistantUsageMessage = {
 	role: "assistant";
 	usage: Record<string, unknown>;
@@ -37,127 +32,19 @@ export default function goalExtension(pi: ExtensionAPI): void {
 	let agentGoalAccounting: AgentGoalAccounting | null = null;
 	let completedThisTurnGoalId: string | null = null;
 
-	pi.registerTool({
-		name: "create_goal",
-		label: "Create Goal",
-		description:
-			"Create a goal only when explicitly requested by the user or system/developer instructions; do not infer goals from ordinary tasks.\nFails if a goal already exists; use update_goal only for status.",
-		parameters: Type.Object(
-			{
-				objective: Type.String({
-					description:
-						"Required. The concrete objective to start pursuing. This starts a new active goal only when no goal is currently defined; if a goal already exists, this tool fails.",
-				}),
-			},
-			{ additionalProperties: false },
-		),
-		async execute(_toolCallId, params, _signal, _onUpdate, ctx) {
-			const ref = goalStoreRef(ctx);
-			if ((await readGoal(ref)) !== null) {
-				throw new Error(
-					"cannot create a new goal because this thread already has a goal; use update_goal only when the existing goal is complete",
-				);
-			}
-			const goal = await createGoal(ref, params.objective);
-			beginAgentGoalAccounting(goal);
-			updateGoalUi(ctx, goal);
-			return toolText(formatGoalToolResponse(goal));
-		},
+	registerGoalTools(pi, {
+		goalStoreRef,
+		accountCurrentAgentTurn,
+		beginAgentGoalAccounting,
+		markGoalCompletedThisTurn,
 	});
-
-	pi.registerTool({
-		name: "update_goal",
-		label: "Update Goal",
-		description:
-			"Update the existing goal.\nUse this tool only to mark the goal achieved.\nSet status to `complete` only when the objective has actually been achieved and no required work remains.\nDo not mark a goal complete merely because you are stopping work.\nYou cannot use this tool to pause or resume a goal; those status changes are controlled by the user or system.\nWhen marking the goal achieved with status `complete`, report the final elapsed time and token usage from the tool result to the user.",
-		parameters: Type.Object(
-			{
-				status: Type.Union(
-					COMPLETABLE_GOAL_STATUS_VALUES.map((status) => Type.Literal(status)),
-					{
-						description:
-							"Required. Set to complete only when the objective is achieved and no required work remains.",
-					},
-				),
-			},
-			{ additionalProperties: false },
-		),
-		async execute(_toolCallId, params, _signal, _onUpdate, ctx) {
-			if (params.status !== "complete") {
-				throw new Error(
-					"update_goal can only mark the existing goal complete; pause and resume are controlled by the user or system",
-				);
-			}
-			await accountCurrentAgentTurn(ctx, EMPTY_USAGE, "active");
-			const goal = await updateGoal(goalStoreRef(ctx), { status: "complete" });
-			markGoalCompletedThisTurn(goal);
-			updateGoalUi(ctx, goal);
-			return toolText(formatGoalToolResponse(goal));
-		},
-	});
-
-	pi.registerTool({
-		name: "get_goal",
-		label: "Get Goal",
-		description: "Get the current goal for this thread, including status, token and elapsed-time usage.",
-		parameters: Type.Object({}, { additionalProperties: false }),
-		async execute(_toolCallId, _params, _signal, _onUpdate, ctx) {
-			const goal = await readGoal(goalStoreRef(ctx));
-			updateGoalUi(ctx, goal);
-			return toolText(formatGoalToolResponse(goal));
-		},
-	});
-
-	pi.registerCommand("goal", {
-		description: "Set, inspect, pause, resume, or clear the persistent goal",
-		handler: async (rawArgs, ctx) => {
-			const command = parseGoalCommand(rawArgs);
-			try {
-				switch (command.kind) {
-					case "show": {
-						const goal = await readGoal(goalStoreRef(ctx));
-						updateGoalUi(ctx, goal);
-						ctx.ui.notify(
-							goal === null ? `${GOAL_USAGE}\n${GOAL_EMPTY_HINT}` : formatGoalForTool(goal),
-							goal ? "info" : "warning",
-						);
-						return;
-					}
-					case "setObjective": {
-						await setGoalObjective(pi, ctx, command.objective);
-						return;
-					}
-					case "setStatus": {
-						if (command.status === "paused") {
-							await accountCurrentAgentTurn(ctx, EMPTY_USAGE, "active");
-						}
-						const goal = await updateGoal(goalStoreRef(ctx), { status: command.status });
-						if (goal.status === "active") {
-							beginAgentGoalAccounting(goal);
-						} else {
-							stopAgentGoalAccounting(goal.id);
-						}
-						updateGoalUi(ctx, goal);
-						ctx.ui.notify(`Goal ${goalStatusLabel(goal.status)}\n${formatGoalForTool(goal)}`, "info");
-						queueGoalContinuation(pi, ctx, goal);
-						return;
-					}
-					case "clear": {
-						await accountCurrentAgentTurn(ctx, EMPTY_USAGE, "active");
-						const cleared = await clearGoal(goalStoreRef(ctx));
-						clearAgentGoalAccounting();
-						updateGoalUi(ctx, null);
-						ctx.ui.notify(
-							cleared ? "Goal cleared" : "No goal to clear\nThis thread does not currently have a goal.",
-							cleared ? "info" : "warning",
-						);
-						return;
-					}
-				}
-			} catch (error) {
-				ctx.ui.notify(error instanceof Error ? error.message : String(error), "error");
-			}
-		},
+	registerGoalCommand(pi, {
+		goalStoreRef,
+		accountCurrentAgentTurn,
+		beginAgentGoalAccounting,
+		stopAgentGoalAccounting,
+		clearAgentGoalAccounting,
+		queueGoalContinuation,
 	});
 
 	pi.on("session_start", async (event, ctx) => {
@@ -198,7 +85,10 @@ export default function goalExtension(pi: ExtensionAPI): void {
 			clearAgentGoalAccounting();
 		}
 		updateGoalUiBestEffort(ctx, goal);
-		if (goal?.status === "active" && shouldQueueGoalContinuationAfterAgentEnd(goal, ctx.hasPendingMessages())) {
+		if (
+			goal?.status === "active" &&
+			shouldQueueGoalContinuationAfterAgentEnd(goal, ctx.hasPendingMessages(), event.messages)
+		) {
 			queueHiddenGoalPrompt(pi, buildContinuationPrompt(goal));
 		}
 	});
@@ -209,33 +99,6 @@ export default function goalExtension(pi: ExtensionAPI): void {
 		}
 		clearAgentGoalAccounting();
 	});
-
-	async function setGoalObjective(pi: ExtensionAPI, ctx: ExtensionContext, objective: string): Promise<void> {
-		const ref = goalStoreRef(ctx);
-		const current = await readGoal(ref);
-		if (current !== null) {
-			const shouldReplace = await confirmReplaceGoal(ctx, objective);
-			if (!shouldReplace) return;
-		}
-
-		if (current?.status === "active") {
-			await accountCurrentAgentTurn(ctx, EMPTY_USAGE, "active");
-		}
-		const goal = current === null ? await createGoal(ref, objective) : await updateGoal(ref, { objective });
-		if (goal.status === "active") beginAgentGoalAccounting(goal);
-		updateGoalUi(ctx, goal);
-		ctx.ui.notify(`Goal ${goalStatusLabel(goal.status)}\n${formatGoalForTool(goal)}`, "info");
-		queueGoalContinuation(pi, ctx, goal);
-	}
-
-	async function confirmReplaceGoal(ctx: ExtensionContext, objective: string): Promise<boolean> {
-		if (!ctx.hasUI) return true;
-		const choice = await ctx.ui.select(`Replace goal?\nNew objective: ${objective}`, [
-			REPLACE_GOAL_CHOICE,
-			CANCEL_REPLACE_GOAL_CHOICE,
-		]);
-		return choice === REPLACE_GOAL_CHOICE;
-	}
 
 	async function maybePromptResumePausedGoal(
 		pi: ExtensionAPI,
@@ -357,10 +220,6 @@ function goalStoreRef(ctx: ExtensionContext): GoalStoreRef {
 
 function cwdStoreKey(cwd: string): string {
 	return createHash("sha256").update(cwd).digest("hex").slice(0, 24);
-}
-
-function toolText(text: string): GoalToolResult {
-	return { content: [{ type: "text" as const, text }], details: {} };
 }
 
 function collectAssistantUsage(messages: unknown[]): TokenUsageSnapshot {
